user

+++ js/allProducts.js
@@ -2,6 +2,24 @@ import { controller } from "./requests.js";
 
 const btnAdd = document.querySelector(".header-btn");
 const products = document.querySelector(".products");
+const btnLogin = document.getElementById('login');
+const addBtn = document.getElementById('addBtn');
+
+controller.isLogin().then(login => {
+  if(login) {
+    addBtn.style.display = 'block';
+    btnLogin.textContent = 'Cerrar sesion';
+    btnLogin.addEventListener('click', e => {
+      controller.sign_Out();
+    });
+  } else{
+    addBtn.style.display = 'none';
+    btnLogin.textContent = 'Login';
+    btnLogin.addEventListener('click', e => {
+      window.location.href = '../html/login.html';
+    });
+  }
+});
 
 const createItem = (data) => {
   const item = document.createElement("div");
@@ -20,23 +38,27 @@ const createItem = (data) => {
   prize.textContent = `$${data.precio}`;
   item.appendChild(prize);
 
-  const edit = document.createElement("img");
-  edit.classList = "edit";
-  edit.src = "../img/edite-img.svg";
-  edit.addEventListener("click", (e) => {
-    window.location.href = `../html/editProduct.html?id=${data.id}`;
-  });
-  item.appendChild(edit);
+  controller.isLogin().then((login) => {
+    if (login) {
+      const edit = document.createElement("img");
+      edit.classList = "edit";
+      edit.src = "../img/edite-img.svg";
+      edit.addEventListener("click", (e) => {
+        window.location.href = `../html/editProduct.html?id=${data.id}`;
+      });
+      item.appendChild(edit);
 
-  const deleteItem = document.createElement("img");
-  deleteItem.classList = "delete";
-  deleteItem.src = "../img/delete-img.svg";
-  deleteItem.addEventListener("click", (e) => {
-    controller.removeItem(data.id).then(res => {
-      window.location.reload();
-    });
+      const deleteItem = document.createElement("img");
+      deleteItem.classList = "delete";
+      deleteItem.src = "../img/delete-img.svg";
+      deleteItem.addEventListener("click", (e) => {
+        controller.removeItem(data.id).then((res) => {
+          window.location.reload();
+        });
+      });
+      item.appendChild(deleteItem);
+    }
   });
-  item.appendChild(deleteItem);
 
   const id = document.createElement("p");
   id.textContent = `#${data.id}`;
@@ -45,11 +67,11 @@ const createItem = (data) => {
   return item;
 };
 
-controller.dataItems().then(info => {
-  Object.keys(info).forEach(item => {
+controller.dataItems().then((info) => {
+  Object.keys(info).forEach((item) => {
     products.appendChild(createItem(info[item]));
-  })
-})
+  });
+});
 
 btnAdd.addEventListener("click", (e) => {
   window.location.href = "../html/addProduct.html";
